Refine fs.SkipDir behavior (and add SkipAll for 7z)
It wasn't entirely clear to me whether fs.SkipDir should cause the remaining entries in a directory to be skipped, if returned on a regular file; or if it should be a no-op in that case.

But the docs say "the named directory," which I suppose implies the named file has to be a directory.

So, remove the "feature" that skips the remainder of the directory if returned on a regular file. This could be confusing for people who just return fs.SkipDir on any entry they want to ignore, and don't want to have to check `if info.IsDir()` all the time.

Now, callers can just return fs.SkipDir any time they want to ignore/skip a file or directory, without care for whether it's a directory or not.

If a caller wants to skip the remainder of the directory a file is in, they can do that logic themselves.